+++ src/common/js/global.js
@@ -2,4 +2,4 @@
  * pwix:layout/src/common/js/global.js
  */
 
-uiLayout = {};
+Layout = {};